pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a2a-society"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk",
    "claude-agent-sdk",
    "fastapi",
    "uvicorn",
    "pydantic>=2.0",
    "click",
    "gitpython",
    "pyyaml",
    "rich",
]

[project.scripts]
hive = "hive.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "respx",
    "ruff",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "B008",    # function call in default argument (FastAPI Depends pattern)
    "RUF006", # store asyncio.create_task reference (test code, not production)
]

[tool.ruff.lint.isort]
known-first-party = ["hive"]
